move source out of top-level
