+++ crossbuild.sh
@@ -1,1 +1,1 @@
-env GOOS=linux GOARCH=amd64 go build -o sq-linux-amd64
+env GOOS=linux GOARCH=amd64 go build -o sq-linux-amd64 cmd/slack-queue/main.go
