+++ pkg/service/remove_action.go
